rangeのベースコンポーネント作成
